tests/wpt/tests/css/css-scroll-anchoring/descend-into-container-with-float.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>

body { height: 4000px; }
#outer { width: 300px; }
#outer:after { content: " "; clear:both; display: table; }
#float {
  float: left; background-color: #ccc;
  height: 500px; width: 100%;
}
#inner { height: 100px; background-color: green; }

</style>
<div id="outer">
  <div id="zeroheight">
    <div id="float">
       <div id="changer"></div>
       <div id="inner"></div>
    </div>
  </div>
</div>
<div>after</div>
<script>

// Tests that we descend into zero-height containers that have floating content.

test(() => {
  document.scrollingElement.scrollTop = 50;
  assert_equals(document.querySelector("#zeroheight").offsetHeight, 0);
  document.querySelector("#changer").style.height = "50px";
  assert_equals(document.scrollingElement.scrollTop, 100);
}, "Zero-height container with float.");

</script>
