LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/descend-into-container-with-float-expected.txt

after

FAIL Zero-height container with float. assert_equals: expected 100 but got 50
